refactor: make onboarding easier
- add tool list to README.md
- change all scripts to have the `env bash` instead of direct `bash`
- add "required env changes" to the `./init.sh` (which can now run
  multiple times
- won't print in `init.sh` if not required
- add doc to `chgm` script
- add to env.source.sample PATH expansion explanation

+++ init.sh
@@ -1,7 +1,13 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 cd $(dirname $0)
 CONFIG_DIR=${HOME}/.config/ocm-container
+
+if [[ -z ${aliases_file} ]]; then
+	echo "run command with aliases file"
+	echo 'aliases_file=$(alias) '"$0 $@"
+	exit
+fi
 
 function init_new_config() {
   if [ ! -f ${CONFIG_DIR}/env.source ]
@@ -44,16 +50,40 @@ else
   init_new_config
 fi
 
-echo "Creating symlink for ocm-container binary (requires sudo permissions...)"
-sudo ln -sfn "$(pwd)/ocm-container.sh" /usr/local/bin/ocm-container
+if ! [[ -L /usr/local/bin/ocm-container ]]; then 
+  echo "Creating symlink for ocm-container binary (requires sudo permissions...)"
+  sudo ln -sfn "$(pwd)/ocm-container.sh" /usr/local/bin/ocm-container
+fi
 
-echo
-echo "Tip: Many developers like to add the following alias:"
-echo "alias ocm-container-stg=\"OCM_URL=staging ocm-container\""
+# ! $( alias ocm-container-stg ) || 
+if [[ $(echo "${aliases_file}" | grep -cw 'ocm-container-stg=') -eq 0 ]]; then
+  echo
+  echo "Tip: Many developers like to add the following alias:"
+  echo "alias ocm-container-stg=\"OCM_URL=staging ocm-container\""
+fi
 
 echo
 echo
 echo "ocm-container configuration can be customized by editing ${CONFIG_DIR}/env.source"
 
-
-
+if [[ $( grep -c '^# REQUIRED:' "${CONFIG_DIR}/env.source") -ne 0 ]]; then
+  echo
+  echo
+  echo "it seems that in '${CONFIG_DIR}/env.source' there are some configurations that are not fufilled"
+  echo "please remove the REQUIRED line once they are set:"
+  echo
+  AWK=$( cat << EOF
+/^# REQUIRED:/
+  {
+    print FILENAME ":" NR, \$0;
+    tmpfs=FS;
+    FS="=";
+    getline;
+    print FILENAME ":" NR, \$1 "=";
+    FS=tmpfs;
+    print "";
+  }
+EOF
+)
+  awk -f <( echo $AWK ) "${CONFIG_DIR}/env.source"
+fi
